% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_pathway_scores.R
\docType{data}
\name{c_pathway_scores}
\alias{c_pathway_scores}
\title{Pathway Scores}
\format{
A matrix of dimension 61x21.
}
\usage{
c_pathway_scores
}
\description{
Pathway scores computed for the C-pathways. Rows correspond to the
61 LGG subjects and the columns correspond to the 21 C-pathways. Row names and
column names indicate the TCGA subject ID and pathway name, respectively.
}
\keyword{datasets}
