% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomPCA.R
\name{geomPCA}
\alias{geomPCA}
\title{Geometric Principal Component Analysis}
\usage{
geomPCA(x, perc.var = 99, m = 512)
}
\arguments{
\item{x}{a list with each entry as a vector of values required to
compute the probability density function for that case/subject}

\item{perc.var}{percentage variance explained by PCs (between 0 to 100)}

\item{m}{number of equally spaced points at which the densities are to be estimated}
}
\value{
pcScores principal component scores; a matrix of size \eqn{n}x\eqn{p},
                 where \eqn{p} is the number of PCs included
}
\description{
Computes the principal component (PC) scores from the geometric principal
component analysis on probability density functions.
}
\examples{
x = lapply(1:10, function(i) rnorm(100))
PCScores = geomPCA(x)
}
