% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupSS.R
\name{groupSS}
\alias{groupSS}
\title{MCMC algorithm for the estimation of the Bayesian model}
\usage{
groupSS(
  y,
  X,
  g_id,
  beta_init = NULL,
  zeta_init = NULL,
  w_init = NULL,
  nusq_init = NULL,
  sigsq_init = 1,
  v0 = 0.005,
  a1 = 0.001,
  a2 = 0.001,
  c1 = 0.001,
  c2 = 0.001,
  Nmcmc = 5000,
  ind = 1:Nmcmc
)
}
\arguments{
\item{y}{response for \eqn{n} subjects}

\item{X}{predictors - dimension \eqn{n} subjects x \eqn{L} predictors}

\item{g_id}{numeric vector indicating group membership in the predictors. Contains values \eqn{1,\ldots,g} if the predictors belong to \eqn{g} groups}

\item{beta_init}{initial values for \eqn{\beta}. Defaults to 0.}

\item{zeta_init}{initial values for \eqn{\zeta}. Defaults to random generation from Ber(0.5).}

\item{w_init}{initial value for \eqn{w} - the complexity parameter. Defaults to proportion of 1s in zeta_init.}

\item{nusq_init}{initial values for \eqn{\nu^2}. Defaults to 1.}

\item{sigsq_init}{initial values for \eqn{\sig^2}. Defaults to 1.}

\item{v0}{hyperparameter \eqn{v0}. Defaults to 0.005.}

\item{a1}{hyperparameter \eqn{a1}. Defaults to 0.001.}

\item{a2}{hyperparameter \eqn{a2}. Defaults to 0.001.}

\item{c1}{hyperparameter \eqn{c1}. Defaults to 0.001.}

\item{c2}{hyperparameter \eqn{c2}. Defaults to 0.001.}

\item{Nmcmc}{number of MCMC samples to generate. Defaults to 5000.}

\item{ind}{indices of MCMC samples to use after burnin and thinning. Defaults to 1 to Nmcmc.}
}
\value{
zeta: MCMC samples of \eqn{\zeta} - dimension \eqn{N} samples x \eqn{g} groups

b: MCMC samples of \eqn{\beta} - dimension \eqn{N} samples x \eqn{L} predictors

x_cnames: column names from the predictor matrix
}
\description{
Gibbs sampling algorithm for the estiamtion of the model with group spike-and-slab prior
}
\examples{
p = 10
n = 100
beta = rep(c(2,0), c(p/2,p/2))
g_id = rep(c(1,2), c(p/2,p/2))
X = matrix(rnorm(n*p), nrow = n)
y = rnorm(n, mean = X\%*\%beta, sd = 0.01)
groupSS(y, X, g_id, Nmcmc = 50)
}
\keyword{groupSS()}
