% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localFDR.R
\name{localFDR}
\alias{localFDR}
\title{Local False Discovery Rates}
\usage{
localFDR(chain, thres = 0.001, FDRc = 0.05)
}
\arguments{
\item{chain}{MCMC samples of the parameter - dimension \eqn{N} samples x \eqn{L} parameters.}

\item{thres}{threshold used to compute the Local FDR. Defaults to 0.001.}

\item{FDRc}{level at which the FDR is controlled (Bayesian FDR). Defaults to 0.05.}
}
\value{
r: local FDRs of the \eqn{L} predictors
}
\description{
Computes the local false discovery rates (FDR) for the paramters given MCMC samples for those parameters
}
\examples{
chain = matrix(rnorm(100*5), nrow=100)
localFDR(chain)
}
\keyword{localFDR()}
