% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\name{Mode}
\alias{Mode}
\title{Mode for a vector of numbers}
\usage{
Mode(x)
}
\arguments{
\item{x}{numeric vector of \eqn{N} MCMC samples (or any numeric vector)}
}
\value{
m mode for the group of numbers
}
\description{
Computes mode for a given set of MCMC samples of a parameter
}
\examples{
Mode()
}
\keyword{Mode()}
