% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_groups.R
\docType{data}
\name{pc_groups}
\alias{pc_groups}
\title{PC Group IDs}
\format{
A vector of length 143.
}
\usage{
pc_groups
}
\description{
Group IDs for the principal component (PC) scores. There 12 groups
formed due to 4 imaging sequences and 3 tumor subregions. Each of these 12 groups
is identified by an integer between 1 to 12. A total of 143 PC scores are included
across the 12 groups. The PC group ID is a vector of size 143 containing integer
values between 1 to 12.
}
\keyword{datasets}
