% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_scores.R
\docType{data}
\name{pc_scores}
\alias{pc_scores}
\title{PC Scores}
\format{
A matrix of dimension 61x143.
}
\usage{
pc_scores
}
\description{
Principal component (PC) scores computed for the 12 groups
formed due to 4 imaging sequences and 3 tumor subregions. Rows correspond
to the 61 LGG subjects and the columns correspond to the 143 PCs. Row names
and column names indicate the TCGA subject ID and PC name, respectively.
}
\keyword{datasets}
