% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pep.R
\name{pep}
\alias{pep}
\title{Posterior Exclusion Probabilities}
\usage{
pep(chain, thres = 0.001, FDRc = 0.05)
}
\arguments{
\item{chain}{MCMC samples of the parameter - dimension \eqn{N} samples x \eqn{L} parameters.}

\item{thres}{threshold used to compute the posterior exclusion probabilities. Defaults to 0.001.}

\item{FDRc}{level at which the FDR is controlled (average Bayesian FDR). Defaults to 0.05.}
}
\value{
r posterior exclusion probabilities of the \eqn{L} predictors
}
\description{
Computes the posterior exclusion probabilities for the paramters given MCMC samples for those parameters
}
\examples{
pep()
}
\keyword{pep()}
