% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr_var_selection.R
\name{fdr_var_selection}
\alias{fdr_var_selection}
\title{FDR-based variable selection}
\usage{
fdr_var_selection(b, x_cnames, ...)
}
\arguments{
\item{b}{MCMC samples of the parameter \eqn{\beta} - dimension \eqn{N} samples x \eqn{L} predictors. Output from the groupSS function}

\item{x_cnames}{names of columns in the predictor matrix}

\item{...}{other arguments to \eqn{localFDR()}}
}
\value{
est_coef: estimated coefficients after the FDR-based variable selection
}
\description{
Computes the estimated coefficients of \eqn{\beta} from the radiogenomics model
}
\examples{
b = matrix(rnorm(100*5), nrow=100)
x_cnames = paste0('Col',1:5)
fdr_var_selection(b,x_cnames)
}
\keyword{fdr_var_selection()}
